refactor: refactoring CANdle and mdtool

+++ MDtool/src/mdtool.hpp
@@ -15,8 +15,8 @@ class Mdtool
 	void ping();
 	bool updateMd80(std::string& filePath, uint32_t id, bool recover, bool all);
 	bool updateBootloader(std::string& filePath, uint32_t id, bool recover);
-	bool readSDO(uint32_t id, uint32_t index, uint32_t subindex);
-	bool writeSDO(uint32_t id, uint32_t index, uint32_t subindex, const IODParser::ValueType& value);
+	bool readSDO(uint32_t id, uint16_t index, uint8_t subindex);
+	bool writeSDO(uint32_t id, uint16_t index, uint8_t subindex, const IODParser::ValueType& value);
 
    private:
 	static constexpr uint32_t secondaryBootloaderAddress = 0x8005000;
